Created cmd for locale to follow DRY principle
