+++ scripts/postinstall.js
@@ -2,11 +2,6 @@
 
 const fs = require('fs');
 const path = require('path');
-const inquirer = require('inquirer');
-const osLocale = require('os-locale');
-// const fileName = path.resolve(__dirname, '..', 'config', 'default.json');
-// const configFile = require(fileName);
-const {countries} = require('../utils');
 
 // Dirty way to bypass TravisCI build process failing due to a timeout
 // waiting for inquirer inputs. Will need to revisit.
@@ -16,50 +11,17 @@ if (process.env.USE_DETECTED_LOCALE) {
 
 if (!fs.existsSync(path.resolve(__dirname, '..', 'config'))) {
   fs.mkdirSync(path.resolve(__dirname, '..', 'config'));
-  const localeSetting = {locale: ""};
+  const localeSetting = {locale: ''};
   fs.writeFileSync(path.resolve(__dirname, '..', 'config', 'default.json'), JSON.stringify(localeSetting, null, 2), {flags: 'as'});
 }
 
 const fileName = path.resolve(__dirname, '..', 'config', 'default.json');
 const configFile = require(fileName);
 
-// Getting the user's system locale...
-let userLocale = osLocale.sync();
+// Setup locale setting
+require('../cmds/locale')(configFile, fileName, displayReadyMessage);
 
-const questions = [
-  {
-    message: `Use the detected system locale (${userLocale}) for AMO queries?`,
-    type: 'confirm',
-    name: 'useSystemLocale',
-    default: true,
-  }, {
-    // Only prompt 'localeSelection' if 'useSystemLocale' returned false
-    when: (response) => {
-      return !response.useSystemLocale;
-    },
-    message: 'What is your preferred locale?',
-    type: 'list',
-    name: 'localeSelection',
-    // incomplete locale list - will be updated as I see more locales
-    // good add-ons to test for correct locale being applied: Ghostery, Ecosia
-    //   Adblock Plus, uBlock Origin
-    choices: countries,
-  },
-];
-
-inquirer.prompt(questions)
-.then((answers) => {
-  userLocale = (answers.localeSelection) ? answers.localeSelection : userLocale;
-  // Setting user locale to the config file
-  configFile.locale = userLocale;
-  fs.writeFile(fileName, JSON.stringify(configFile, null, 2), (err) => {
-    if (err) throw err;
-    console.log(`Successfully set the user locale as: ${userLocale}`);
-    displayReadyMessage();
-  });
-});
-
-const displayReadyMessage = () => {
-  // You're all set. Use `amo help` to learn how to get started. Enjoy!
-  // If you need to change your locale, run `amo locale`.
-};
+function displayReadyMessage() {
+  console.log(`You're all set. Use 'amo help' to learn how to get started. Enjoy!
+  If you would like to change your locale, run 'amo locale'.`);
+}
